refactor(nx-plugin): ship ESM-only via stock @nx/js:swc executor
Drop the CommonJS build entirely and produce a single ESM build using the
standard `@nx/js:swc` executor instead of hand-rolled swc CLI commands and a
custom finalize-package.mjs.

Rationale:
- Node unflagged `require(esm)` for synchronous ESM in 20.19.0 / 22.12.0, so
  CommonJS consumers can `require('@aws/nx-plugin/...')` against the ESM build
  directly (the SDK has no top-level await). Nx 23 loads generators as ESM via
  native type stripping. Neither path needs a CJS build.
- The only reason the dual build used custom swc commands was converting
  `import.meta.dirname` -> `__dirname` for CJS output. With CJS gone, the stock
  executor (swc + tsc declarations + asset/package.json copy) does everything.

Changes:
- Move `sdk/` and `bin/` under `src/` so the project has a single source root
  the `@nx/js:swc` executor compiles in one pass (it can't take multiple roots,
  and `additionalEntryPoints` is unsupported under the TS-solution setup).
  `exports`/`bin` now point at `./src/sdk/*` / `./src/bin/*`; internal
  `../src/...` imports become `../...`.
- `package.json`: ESM-only `exports` (`types`+`default`, no `require`/`cjs`),
  add `engines.node >= 20.19.0`.
- Delete `.swcrc.cjs` and `scripts/finalize-package.mjs`; no more `cjs/` dir or
  `type: commonjs` marker.
- `ts#nx-generator` `addSdkExport` writes to `src/sdk/*` with `../<dir>` module
  specifiers.

Verified via verdaccio: ESM `import` and CJS `require()` of every SDK subpath
work; the vended `.ts` generator loads under native strip and its plugin
`package` target builds.
