+++ packages/nx-plugin/.swcrc
@@ -4,13 +4,25 @@
   "jsc": {
     "target": "es2022",
     "baseUrl": "src",
-    "parser": { "syntax": "typescript", "decorators": true },
-    "transform": { "legacyDecorator": true, "decoratorMetadata": true },
+    "parser": {
+      "syntax": "typescript",
+      "decorators": true
+    },
+    "transform": {
+      "legacyDecorator": true,
+      "decoratorMetadata": true
+    },
     "keepClassNames": true,
     "externalHelpers": false,
-    "experimental": { "keepImportAttributes": true }
+    "experimental": {
+      "keepImportAttributes": true
+    }
   },
-  "module": { "type": "es6", "resolveFully": true, "strict": true },
+  "module": {
+    "type": "es6",
+    "resolveFully": true,
+    "strict": true
+  },
   "exclude": [
     ".*\\.spec\\.ts$",
     ".*\\.test\\.ts$",
