+++ packages/nx-plugin/package.json
@@ -9,20 +9,18 @@
   "private": false,
   "type": "module",
   "license": "Apache-2.0",
-  "main": "./cjs/src/index.js",
+  "main": "./src/index.js",
   "module": "./src/index.js",
   "types": "./src/index.d.ts",
   "typings": "./src/index.d.ts",
   "exports": {
     ".": {
       "types": "./src/index.d.ts",
-      "import": "./src/index.js",
-      "require": "./cjs/src/index.js"
+      "default": "./src/index.js"
     },
     "./sdk/*": {
-      "types": "./sdk/*.d.ts",
-      "import": "./sdk/*.js",
-      "require": "./cjs/sdk/*.js"
+      "types": "./src/sdk/*.d.ts",
+      "default": "./src/sdk/*.js"
     },
     "./package.json": "./package.json"
   },
@@ -30,10 +28,13 @@
     "access": "public"
   },
   "bin": {
-    "aws-nx-mcp": "./bin/aws-nx-mcp.js"
+    "aws-nx-mcp": "./src/bin/aws-nx-mcp.js"
   },
   "generators": "./generators.json",
   "executors": "./executors.json",
+  "engines": {
+    "node": ">=20.19.0"
+  },
   "peerDependencies": {
     "nx": "23.0.1"
   },
