+++ packages/nx-plugin/project.json
@@ -6,25 +6,50 @@
   "tags": [],
   "targets": {
     "compile": {
-      "executor": "nx:run-commands",
-      "outputs": ["{workspaceRoot}/dist/packages/nx-plugin"],
-      "cache": true,
-      "dependsOn": ["^compile"],
+      "executor": "@nx/js:swc",
+      "outputs": [
+        "{options.outputPath}"
+      ],
       "options": {
-        "cwd": "{projectRoot}",
-        "parallel": false,
-        "commands": [
-          "node ../../node_modules/@swc/cli/bin/swc.js src sdk bin -d ../../dist/packages/nx-plugin --config-file=.swcrc --copy-files",
-          "node ../../node_modules/@swc/cli/bin/swc.js src sdk bin -d ../../dist/packages/nx-plugin/cjs --config-file=.swcrc.cjs",
-          "tsc -p tsconfig.lib.json --declaration --emitDeclarationOnly --rootDir . --outDir ../../dist/packages/nx-plugin",
-          "node scripts/finalize-package.mjs"
+        "outputPath": "dist/packages/nx-plugin",
+        "main": "packages/nx-plugin/src/index.ts",
+        "tsConfig": "packages/nx-plugin/tsconfig.lib.json",
+        "swcrc": "packages/nx-plugin/.swcrc",
+        "stripLeadingPaths": false,
+        "assets": [
+          "packages/nx-plugin/*.md",
+          "packages/nx-plugin/LICENSE*",
+          "packages/nx-plugin/NOTICE",
+          {
+            "input": "./packages/nx-plugin/src",
+            "glob": "**/!(*.ts)",
+            "output": "./src"
+          },
+          {
+            "input": "./packages/nx-plugin/src",
+            "glob": "**/*.d.ts",
+            "output": "./src"
+          },
+          {
+            "input": "./packages/nx-plugin",
+            "glob": "generators.json",
+            "output": "."
+          },
+          {
+            "input": "./packages/nx-plugin",
+            "glob": "executors.json",
+            "output": "."
+          }
         ]
       }
     },
     "lint": {
       "executor": "nx:run-commands",
       "cache": true,
-      "inputs": ["default", "{workspaceRoot}/biome.json"],
+      "inputs": [
+        "default",
+        "{workspaceRoot}/biome.json"
+      ],
       "options": {
         "command": "biome lint {projectRoot}/src"
       },
@@ -47,12 +72,20 @@
     "post-compile": {
       "executor": "nx:run-commands",
       "options": {
-        "commands": ["chmod +x dist/packages/nx-plugin/bin/aws-nx-mcp.js"]
+        "commands": [
+          "chmod +x dist/packages/nx-plugin/src/bin/aws-nx-mcp.js"
+        ]
       },
-      "dependsOn": ["compile"]
+      "dependsOn": [
+        "compile"
+      ]
     },
     "package": {
-      "dependsOn": ["compile", "post-compile", "generate-3p-license"]
+      "dependsOn": [
+        "compile",
+        "post-compile",
+        "generate-3p-license"
+      ]
     },
     "build": {
       "dependsOn": [
@@ -66,7 +99,9 @@
     "mcp-inspect": {
       "executor": "nx:run-commands",
       "options": {
-        "commands": ["mcp-inspector -- tsx --watch bin/aws-nx-mcp.ts"],
+        "commands": [
+          "mcp-inspector -- tsx --watch src/bin/aws-nx-mcp.ts"
+        ],
         "cwd": "{projectRoot}"
       }
     }
