--- packages/nx-plugin/sdk/agentcore-gateway.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-// AgentCore Gateway Generator
-export { agentcoreGatewayGenerator } from '../src/agentcore-gateway/generator';
-export type { AgentcoreGatewayGeneratorSchema } from '../src/agentcore-gateway/schema';
