--- packages/nx-plugin/sdk/connection.ts
@@ -1,7 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { connectionGenerator } from '../src/connection/generator';
-export type { ConnectionGeneratorSchema } from '../src/connection/schema';
