--- packages/nx-plugin/sdk/open-api.ts
@@ -1,16 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export {
-  openApiTsClientGenerator,
-  buildOpenApiCodeGenerationData,
-} from '../src/open-api/ts-client/generator';
-export type { OpenApiTsClientGeneratorSchema } from '../src/open-api/ts-client/schema';
-
-export { openApiTsHooksGenerator } from '../src/open-api/ts-hooks/generator';
-export type { OpenApiTsHooksGeneratorSchema } from '../src/open-api/ts-hooks/schema';
-
-export { openApiTsMetadataGenerator } from '../src/open-api/ts-metadata/generator';
-export type { OpenApiTsMetadataGeneratorSchema } from '../src/open-api/ts-metadata/schema';
