--- packages/nx-plugin/sdk/py.ts
@@ -1,18 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { pyAgentGenerator } from '../src/py/agent/generator';
-export type { PyAgentGeneratorSchema } from '../src/py/agent/schema';
-
-export { pyApiGenerator } from '../src/py/api/generator';
-export type { PyApiGeneratorSchema } from '../src/py/api/schema';
-
-export { pyLambdaFunctionGenerator } from '../src/py/lambda-function/generator';
-export type { PyLambdaFunctionGeneratorSchema } from '../src/py/lambda-function/schema';
-
-export { pyMcpServerGenerator } from '../src/py/mcp-server/generator';
-export type { PyMcpServerGeneratorSchema } from '../src/py/mcp-server/schema';
-export { pyProjectGenerator } from '../src/py/project/generator';
-export type { PyProjectGeneratorSchema } from '../src/py/project/schema';
