--- packages/nx-plugin/sdk/terraform.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-// Terraform Project Generator
-export { terraformProjectGenerator } from '../src/terraform/project/generator';
-export type { TerraformProjectGeneratorSchema } from '../src/terraform/project/schema';
