+++ packages/nx-plugin/src/bin/aws-nx-mcp.ts
@@ -3,7 +3,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { startMcpServer } from '../src/mcp-server';
+import { startMcpServer } from '../mcp-server';
 
 void (async () => {
   try {
