+++ packages/nx-plugin/src/sdk/agentcore-gateway.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+// AgentCore Gateway Generator
+export { agentcoreGatewayGenerator } from '../agentcore-gateway/generator';
+export type { AgentcoreGatewayGeneratorSchema } from '../agentcore-gateway/schema';
