+++ packages/nx-plugin/src/sdk/connection.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { connectionGenerator } from '../connection/generator';
+export type { ConnectionGeneratorSchema } from '../connection/schema';
