+++ packages/nx-plugin/src/sdk/license.ts
@@ -2,17 +2,17 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-export { licenseGenerator } from '../src/license/generator';
+export { licenseGenerator } from '../license/generator';
 export type {
   AllowlistEntry,
   CollectedDependency,
   DependencyCheckConfig,
   DependencyCheckException,
   LicenseCollector,
-} from '../src/license/index';
+} from '../license/index';
 export {
   DEFAULT_LICENSE_ALLOWLIST,
   npmCollector,
   pythonCollector,
-} from '../src/license/index';
-export type { LicenseGeneratorSchema } from '../src/license/schema';
+} from '../license/index';
+export type { LicenseGeneratorSchema } from '../license/schema';
