+++ packages/nx-plugin/src/sdk/open-api.ts
@@ -1,0 +1,16 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export {
+  openApiTsClientGenerator,
+  buildOpenApiCodeGenerationData,
+} from '../open-api/ts-client/generator';
+export type { OpenApiTsClientGeneratorSchema } from '../open-api/ts-client/schema';
+
+export { openApiTsHooksGenerator } from '../open-api/ts-hooks/generator';
+export type { OpenApiTsHooksGeneratorSchema } from '../open-api/ts-hooks/schema';
+
+export { openApiTsMetadataGenerator } from '../open-api/ts-metadata/generator';
+export type { OpenApiTsMetadataGeneratorSchema } from '../open-api/ts-metadata/schema';
