+++ packages/nx-plugin/src/sdk/terraform.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+// Terraform Project Generator
+export { terraformProjectGenerator } from '../terraform/project/generator';
+export type { TerraformProjectGeneratorSchema } from '../terraform/project/schema';
