+++ packages/nx-plugin/src/sdk/ts.ts
@@ -1,0 +1,50 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+// TypeScript Infrastructure
+export { tsInfraGenerator } from '../infra/app/generator';
+export type { TsInfraGeneratorSchema } from '../infra/app/schema';
+// TypeScript Agent Generator
+export { tsAgentGenerator } from '../ts/agent/generator';
+export type { TsAgentGeneratorSchema } from '../ts/agent/schema';
+// TypeScript API
+export { tsApiGenerator } from '../ts/api/generator';
+export type { TsApiGeneratorSchema } from '../ts/api/schema';
+// Documentation Site Generator
+export { tsDocsGenerator } from '../ts/docs/generator';
+export type { TsDocsGeneratorSchema } from '../ts/docs/schema';
+// DynamoDB Generator
+export { tsDynamoDBGenerator } from '../ts/dynamodb/generator';
+export type { TsDynamoDBGeneratorSchema } from '../ts/dynamodb/schema';
+// TypeScript Lambda Function
+export { tsLambdaFunctionGenerator } from '../ts/lambda-function/generator';
+export type { TsLambdaFunctionGeneratorSchema } from '../ts/lambda-function/schema';
+// TypeScript Project Generator
+export { tsProjectGenerator } from '../ts/lib/generator';
+export type { TsProjectGeneratorSchema } from '../ts/lib/schema';
+// TypeScript MCP Server Generator
+export { tsMcpServerGenerator } from '../ts/mcp-server/generator';
+export type { TsMcpServerGeneratorSchema } from '../ts/mcp-server/schema';
+// TypeScript Nx Generator Generator
+export { tsNxGeneratorGenerator } from '../ts/nx-generator/generator';
+export type { TsNxGeneratorGeneratorSchema } from '../ts/nx-generator/schema';
+// TypeScript Nx Plugin Generator
+export { tsNxPluginGenerator } from '../ts/nx-plugin/generator';
+export type { TsNxPluginGeneratorSchema } from '../ts/nx-plugin/schema';
+// Relational Database Generator
+export { tsRdbGenerator } from '../ts/rdb/generator';
+export type { TsRdbGeneratorSchema } from '../ts/rdb/schema';
+// Runtime Config
+export { runtimeConfigGenerator } from '../ts/react-website/runtime-config/generator';
+export type { RuntimeConfigGeneratorSchema } from '../ts/react-website/runtime-config/schema';
+// TypeScript Website Generator
+export { tsWebsiteGenerator } from '../ts/website/app/generator';
+export type { TsWebsiteGeneratorSchema } from '../ts/website/app/schema';
+// TypeScript Website Auth Generator
+export { tsWebsiteAuthGenerator } from '../ts/website/auth/generator';
+export type { TsWebsiteAuthGeneratorSchema } from '../ts/website/auth/schema';
+
+// Shared Constructs
+export { sharedConstructsGenerator } from '../utils/shared-constructs';
