+++ packages/nx-plugin/src/sdk/utils/ast.ts
@@ -2,5 +2,5 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-export { applyGritQL } from '../../src/utils/ast';
-export { matchGritQL } from '../../src/utils/ast';
+export { applyGritQL } from '../../utils/ast';
+export { matchGritQL } from '../../utils/ast';
