+++ packages/nx-plugin/src/sdk/utils/test.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { formatFilesInSubtree } from '../../src/utils/format';
+export { createTreeUsingTsSolutionSetup } from '../../utils/test';
