+++ packages/nx-plugin/src/ts/nx-generator/generator.spec.ts
@@ -303,8 +303,8 @@ describe('nx-generator generator', () => {
 
     it('should add a ts# generator to the ts sdk', async () => {
       tree.write(
-        'packages/nx-plugin/sdk/ts.ts',
-        `export { tsProjectGenerator } from '../src/ts/lib/generator';\n`,
+        'packages/nx-plugin/src/sdk/ts.ts',
+        `export { tsProjectGenerator } from '../ts/lib/generator';\n`,
       );
 
       await tsNxGeneratorGenerator(tree, {
@@ -313,21 +313,21 @@ describe('nx-generator generator', () => {
         description: 'Some description',
       });
 
-      const sdk = tree.read('packages/nx-plugin/sdk/ts.ts', 'utf-8');
+      const sdk = tree.read('packages/nx-plugin/src/sdk/ts.ts', 'utf-8');
       expect(sdk).toContain(
-        `export { tsCoolThingGenerator } from '../src/ts-cool-thing/generator';`,
+        `export { tsCoolThingGenerator } from '../ts-cool-thing/generator';`,
       );
       expect(sdk).toContain(
-        `export type { TsCoolThingGeneratorSchema } from '../src/ts-cool-thing/schema';`,
+        `export type { TsCoolThingGeneratorSchema } from '../ts-cool-thing/schema';`,
       );
       // Existing exports preserved
       expect(sdk).toContain(
-        `export { tsProjectGenerator } from '../src/ts/lib/generator';`,
+        `export { tsProjectGenerator } from '../ts/lib/generator';`,
       );
     });
 
     it('should add a py# generator to the py sdk', async () => {
-      tree.write('packages/nx-plugin/sdk/py.ts', '');
+      tree.write('packages/nx-plugin/src/sdk/py.ts', '');
 
       await tsNxGeneratorGenerator(tree, {
         project: NxPluginForAwsProjectJson.name,
@@ -335,17 +335,17 @@ describe('nx-generator generator', () => {
         description: 'Some description',
       });
 
-      const sdk = tree.read('packages/nx-plugin/sdk/py.ts', 'utf-8');
+      const sdk = tree.read('packages/nx-plugin/src/sdk/py.ts', 'utf-8');
       expect(sdk).toContain(
-        `export { pyCoolThingGenerator } from '../src/py-cool-thing/generator';`,
+        `export { pyCoolThingGenerator } from '../py-cool-thing/generator';`,
       );
       expect(sdk).toContain(
-        `export type { PyCoolThingGeneratorSchema } from '../src/py-cool-thing/schema';`,
+        `export type { PyCoolThingGeneratorSchema } from '../py-cool-thing/schema';`,
       );
     });
 
     it('should add a ts# generator to the ts sdk in a nested directory', async () => {
-      tree.write('packages/nx-plugin/sdk/ts.ts', '');
+      tree.write('packages/nx-plugin/src/sdk/ts.ts', '');
 
       await tsNxGeneratorGenerator(tree, {
         project: NxPluginForAwsProjectJson.name,
@@ -354,17 +354,17 @@ describe('nx-generator generator', () => {
         description: 'Some description',
       });
 
-      const sdk = tree.read('packages/nx-plugin/sdk/ts.ts', 'utf-8');
+      const sdk = tree.read('packages/nx-plugin/src/sdk/ts.ts', 'utf-8');
       expect(sdk).toContain(
-        `export { tsCoolThingGenerator } from '../src/nested/dir/generator';`,
+        `export { tsCoolThingGenerator } from '../nested/dir/generator';`,
       );
       expect(sdk).toContain(
-        `export type { TsCoolThingGeneratorSchema } from '../src/nested/dir/schema';`,
+        `export type { TsCoolThingGeneratorSchema } from '../nested/dir/schema';`,
       );
     });
 
     it('should not duplicate the sdk export on a same-name re-run', async () => {
-      tree.write('packages/nx-plugin/sdk/ts.ts', '');
+      tree.write('packages/nx-plugin/src/sdk/ts.ts', '');
 
       const options = {
         project: NxPluginForAwsProjectJson.name,
@@ -374,15 +374,15 @@ describe('nx-generator generator', () => {
       await tsNxGeneratorGenerator(tree, options);
       await tsNxGeneratorGenerator(tree, options);
 
-      const sdk = tree.read('packages/nx-plugin/sdk/ts.ts', 'utf-8');
+      const sdk = tree.read('packages/nx-plugin/src/sdk/ts.ts', 'utf-8');
       const occurrences =
-        sdk.split(`from '../src/ts-cool-thing/generator';`).length - 1;
+        sdk.split(`from '../ts-cool-thing/generator';`).length - 1;
       expect(occurrences).toBe(1);
     });
 
     it('should not add an sdk export for generators without a known prefix', async () => {
-      tree.write('packages/nx-plugin/sdk/ts.ts', '');
-      tree.write('packages/nx-plugin/sdk/py.ts', '');
+      tree.write('packages/nx-plugin/src/sdk/ts.ts', '');
+      tree.write('packages/nx-plugin/src/sdk/py.ts', '');
 
       await tsNxGeneratorGenerator(tree, {
         project: NxPluginForAwsProjectJson.name,
@@ -390,8 +390,8 @@ describe('nx-generator generator', () => {
         description: 'Some description',
       });
 
-      expect(tree.read('packages/nx-plugin/sdk/ts.ts', 'utf-8')).toBe('');
-      expect(tree.read('packages/nx-plugin/sdk/py.ts', 'utf-8')).toBe('');
+      expect(tree.read('packages/nx-plugin/src/sdk/ts.ts', 'utf-8')).toBe('');
+      expect(tree.read('packages/nx-plugin/src/sdk/py.ts', 'utf-8')).toBe('');
     });
 
     it('should throw if given a project that is not @aws/nx-plugin', async () => {
